+++ JacobiOMP.cpp
@@ -23,23 +23,23 @@ using namespace std;
 void JacobiOMP::freeAllMemory() {
 	freeMemory();
 }
-void JacobiOMP::solve(double eps, int num_threads) {
+void JacobiOMP::solve(float eps, int num_threads) {
     cout << "number of threads: "<<num_threads<<std::endl;
     omp_set_num_threads(num_threads);
     JacobiOMP::solve(eps);
 
 }
 
-void JacobiOMP::solve(double eps) {
+void JacobiOMP::solve(float eps) {
 
 	high_resolution_clock::time_point start;
     high_resolution_clock::time_point end;
-    duration<double, std::milli> duration_sec;
+    duration<float, std::milli> duration_sec;
     
-	//double residual = 0.0;  //	
-	//double sum = 0.0;
-	// double dis = 0.0;
-	// double diff = 1.0;  
+	//float residual = 0.0;  //	
+	//float sum = 0.0;
+	// float dis = 0.0;
+	// float diff = 1.0;  
 	// int multicity = int(0.1 / eps);
 	//timer.start();
 	// Get the starting timestamp
@@ -58,7 +58,7 @@ void JacobiOMP::solve(double eps) {
         //#pragma omp parallel for collapse(2) reduction(+:sum)
 		for (int i = 0; i < size; i++)
 		{
-            double sum = 0.0;
+            float sum = 0.0;
             //#pragma omp parallel for
             //#pragma omp parallel for reduction(+:sum)
 			for (int j = 0; j < size; j++)
@@ -86,28 +86,28 @@ void JacobiOMP::solve(double eps) {
 		// if (diff < eps*multicity) {
 		// 	multicity = int(multicity / 10);
 		// }
-		memcpy(x, nextX, size * sizeof(double));
+		memcpy(x, nextX, size * sizeof(float));
 	}
 	// Get the ending timestamp
 	end = high_resolution_clock::now();
 	cout << endl << "Iterations:" << count << endl;
 	
     
-    // Convert the calculated duration to a double using the standard library
-    duration_sec = std::chrono::duration_cast<duration<double, std::milli>>(end - start);
+    // Convert the calculated duration to a float using the standard library
+    duration_sec = std::chrono::duration_cast<duration<float, std::milli>>(end - start);
 	cout << duration_sec.count() << "\n";
 
 }
-// void JacobiOMP::solve(double eps) {
+// void JacobiOMP::solve(float eps) {
 
 // 	high_resolution_clock::time_point start;
 //     high_resolution_clock::time_point end;
-//     duration<double, std::milli> duration_sec;
+//     duration<float, std::milli> duration_sec;
     
-// 	double residual = 0.0;  //	
-// 	//double sum = 0.0;
-// 	double dis = 0.0;
-// 	double diff = 1.0;  
+// 	float residual = 0.0;  //	
+// 	//float sum = 0.0;
+// 	float dis = 0.0;
+// 	float diff = 1.0;  
 // 	int multicity = int(0.1 / eps);
 // 	//timer.start();
 // 	// Get the starting timestamp
@@ -126,7 +126,7 @@ void JacobiOMP::solve(double eps) {
 //         //#pragma omp parallel for collapse(2) reduction(+:sum)
 // 		for (int i = 0; i < size; i++)
 // 		{
-//             double sum = 0.0;
+//             float sum = 0.0;
 //             //#pragma omp parallel for
 //             //#pragma omp parallel for reduction(+:sum)
 // 			for (int j = 0; j < size; j=j+2)
@@ -147,15 +147,15 @@ void JacobiOMP::solve(double eps) {
 // 		if (diff < eps*multicity) {
 // 			multicity = int(multicity / 10);
 // 		}
-// 		memcpy(x, nextX, size * sizeof(double));
+// 		memcpy(x, nextX, size * sizeof(float));
 // 	}
 // 	// Get the ending timestamp
 // 	end = high_resolution_clock::now();
 // 	cout << endl << "Iterations:" << count << endl;
 	
     
-//     // Convert the calculated duration to a double using the standard library
-//     duration_sec = std::chrono::duration_cast<duration<double, std::milli>>(end - start);
+//     // Convert the calculated duration to a float using the standard library
+//     duration_sec = std::chrono::duration_cast<duration<float, std::milli>>(end - start);
 // 	cout << duration_sec.count() << "\n";
 
 // }
@@ -207,7 +207,7 @@ int main(int argc, char ** argv) {
 	//cout << dimension;
 	JacobiOMP * jacobi = new JacobiOMP(dimension);
 	jacobi->input(argv[2], generate_random);
-	double eps = stod(argv[4]);
+	float eps = stod(argv[4]);
 	jacobi->solve(eps, num_threads);
 	jacobi->output(argv[5]);
 	jacobi->computeError();
